Fixed saving bug Bug trait points in UI are not saving

+++ Assets/Scripts/Inventory/Purse.cs
@@ -28,7 +28,7 @@ namespace RPG.Inventories
 
         public void RestoreState(object state)
         {
-            _currency = (int) state;
+            _currency = (float) state;
         }
     }
 }
